requirements.txt
chardet~=3.0.4
### BEGIN ###
scrapy>=2.4.0
SQLAlchemy>=1.2.16
mysql-connector-python>=8.0.14
pyexcel>=0.6.6
requests~=2.24.0
google-api-python-client~=1.12.8
google-auth-httplib2~=0.0.4
google-auth-oauthlib~=0.4.2
pyodbc~=4.0.30
oauth2client~=4.1.3
sphinx~=3.5.4
sphinx-rtd-theme~=0.5.2
### END ###